requirements.txt
requests>=2.25.0
pydantic>=1.8.0
Pillow>=8.0.0
torch>=1.9.0